requirements.txt
Django==3.1
djangorestframework==3.11.1
psycopg2-binary==2.8.5
pytest-django==3.9.0